Create number column component
