--- src/components/Counter/NumberColumn.jsx
@@ -1,60 +1,0 @@
-import React, { useEffect, useRef, useState } from 'react'
-import {motion} from 'framer-motion'
-
-export default function NumberColumn({ digit }) {
-   const [pos, setPos] = useState(0);
-   const container = useRef();
-
-   const digitsArray = Array.from(Array(10).keys()).reverse(); 
-
-   const setColumnToNumber = (number) => {
-      setPos(container.current.clientHeight * parseInt(number,10));
-   }
-
-   useEffect(() => setColumnToNumber(digit), [digit]);
-
-   const Decimal = (d) => {
-      return(
-         <div className="counterDot">
-            <div className="digit">
-               <span>{d}</span>
-            </div>
-         </div>
-      )
-   }
-
-   const digitAnim ={
-      anim: {
-         y: pos
-      }
-   }
-
-   const digitCont = {
-      show: {
-         transition: {
-           staggerChildren: 1
-         }
-       }
-   }
-
-   const Digit = () => {
-      return(
-         <>
-            <motion.div variants={digitAnim} animate="anim" className="counterColumn">
-               {digitsArray.map(no => (
-                  <div key={`n${no}`} className="digit">
-                     <span>{no}</span>
-                  </div>
-               ))}
-            </motion.div>
-            <span className="digitPlaceholder">0</span>
-         </>
-      )
-   }
-
-   return (
-      <motion.div variants={digitCont} animate="show" className="counterContainer" ref={container}>
-         {digit.match(/\D+/) ? Decimal(digit):Digit()}
-      </motion.div>
-   )
-}
