+++ src/components/Counter/NumberColumn/NumberColumn.anim.jsx
@@ -1,0 +1,13 @@
+export const digitAnim = {
+	anim: (custom) => ({
+		y: custom,
+	}),
+};
+
+export const digitCont = {
+	show: {
+		transition: {
+			staggerChildren: 1,
+		},
+	},
+};
